fix: core structure
