+++ tools/memory/lib/core/PackageBuilder.js
@@ -7,6 +7,7 @@
  * @author AXIVO
  * @license BSD-3-Clause
  */
+
 const fs = require('fs');
 const path = require('path');
