+++ tools/memory/lib/loaders/ConfigLoader.js
@@ -7,9 +7,10 @@
  * @author AXIVO
  * @license BSD-3-Clause
  */
+
 const fs = require('fs');
-const path = require('path');
 const os = require('os');
+const path = require('path');
 const yaml = require('js-yaml');
 const MemoryBuilderError = require('../core/Error');
